feat(desktop): reuse the extension's animated loading screen
Replace the static "Loading session data" spinner with the extension's
"Building Activity Index" loading card — animated progress bar, step
checklist, live file counts and elapsed timer.

- extract the loading screen builders (CSS, body, progress script) from
  extension.ts into a shared vscode-free module, loadingHtml.ts; the
  extension delegates to it so both surfaces stay identical
- desktop composes the same pieces in buildLoadingHtml, embedding the app
  icon as a data URI
- preload bridges 'loading-message' IPC to window.postMessage so the shared
  script receives progress exactly like a VS Code webview
- getStats now passes calculateDetailedStats' existing progressCallback,
  throttled to 500ms like the extension, emitting discovering/parsing/
  computing steps and live percentages
- fix the dark-theme --vscode-descriptionForeground seed (#717171 was the
  light-theme value) to #9d9d9d so muted text like the loading subtitle is
  readable, matching what VS Code's dark theme actually renders

Verified in the running app: live progress, step transitions, counters and
elapsed timer all animate; subtitle contrast now ~6.5:1 (was ~2.9:1).
Extension unit tests pass (80 files).
